api/requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.0
Flask==2.0.2
Flask-HTTPAuth==4.5.0
Flask-SQLAlchemy==2.5.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
